azure-requirements.txt
concurrent-log-handler
opencv-python
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
mlflow
azureml-mlflow
